+++ test/src/analyzers/lint_analyzer/rules/rules_list/prefer_using_list_view/examples/example.dart
@@ -1,0 +1,67 @@
+import 'package:flutter/material.dart';
+
+void main() {
+  runApp(MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    // LINT
+    return SingleChildScrollView(
+      child: Column(
+        children: [
+          Text('Wow lint rule'),
+          Text('Wow another lint rule'),
+        ],
+      ),
+    );
+  }
+}
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return SingleChildScrollView(
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text('Wow lint rule'),
+          Text('Wow another lint rule'),
+        ],
+      ),
+    );
+  }
+}
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return SingleChildScrollView(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.start,
+        children: [
+          Text('Wow lint rule'),
+          Text('Wow another lint rule'),
+        ],
+      ),
+    );
+  }
+}
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: Scaffold(
+        body: SingleChildScrollView(
+          child: Text(),
+        ),
+      ),
+    );
+  }
+}
+
+class SingleChildScrollView {}
+
+class Column {}