+++ test/src/analyzers/lint_analyzer/rules/rules_list/prefer_conditional_expressions/examples/example.dart
@@ -173,4 +173,12 @@ int newCase() {
   } else {
     value /= 5;
   }
+
+  // LINT
+  bool val = false;
+  if (true) {
+    val = true;
+  } else {
+    val = false;
+  }
 }